reorganized operator controls components
