+++ src/components/controls/plan/OperatorLevel.vue
