+++ src/components/controls/plan/OperatorSkillLevels.vue
