+++ src/components/controls/plan/OperatorSkillMasteries.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed, ref } from 'vue';
-import OperatorInputGroup from '@/components/controls/OperatorInputGroup.vue';
+import OperatorInputGroup from '@/components/controls/plan/OperatorInputGroup.vue';
 import { SelectedOperator } from '@/types/planner-types';
 
 const props = defineProps<{
